# measure test performance time