--- a/dev/test.mjs
+++ b/dev/test.mjs
@@ -1,20 +1,25 @@
 import { loadModuleMap } from './module.mjs'
+import { performance } from 'node:perf_hooks'
+
+/** @type {(s: string) => <T>(_: T) => T} */
+const log = s => state => {
+    console.log(s)
+    return state
+}
+
+/** @type {<T>(_: T) => readonly[number, T]} */
+const performanceNow = state => [performance.now(), state]
 
 // test runner.
 const main = async() => {
     const moduleMap = await loadModuleMap()
-
-    /** @type {(s: string) => <T>(_: T) => T} */
-    const log = s => state => {
-        console.log(s)
-        return state
-    }
 
     /** @type {any} */
     const f = moduleMap['./dev/test/module.f.cjs'].exports
     f({
         moduleMap,
         log,
+        performanceNow,
     })
 }