- Deno (1.29.1): 614.1098729999949 ms
- Node (19): 796.0818229999568 ms
- Node (18): 843.4700999999186 ms
- Bun: 936.2019029999988 ms
- Node (16): 947.3171790001215 ms